Split parser module atoms
